requirements.txt
pillow
opencv-python-headless
numpy
pyzbar
pyaadhaar
xmltodict
pytesseract
psutil
deepface